test(webhook): mirror check_runs split in test files
Move tests next to the production files they exercise after the
previous check_runs.go split:

- check_aggregate_test.go: aggregate, IsAggregateCheck,
  AggregateSummary, ConclusionEmoji, AggregateCheckNameForEnv,
  FilterChecksByEnvironment
- check_records_test.go: LatestApplyByCheckKey
- check_prior_env_test.go: CheckPriorEnvViaLocal* +
  CheckPriorEnvViaGitHub
- check_runs_test.go: shared storage helpers + residual
  webhook-environment / respond-to-unscoped tests

No test logic changes.

Validation:
- go test -count=1 -race ./pkg/webhook/...
- golangci-lint run --new-from-rev=origin/main ./pkg/webhook/... (0 issues)
